requirements.txt
pyrogram-2.0.102
TgCrypto
spotdl
requests
yt-dlp